requirements.txt
langchain-google-genai 
fastapi
uvicorn
python-dotenv
langchain